[usm] service discovery, only non-linux in targetenvs pkg, add target envs SPRING_...
